Add PM2.5 support for IconS devices
IconS devices don't report AQI or PM10, but do report PM2.5 both on the
device's display and via the API. Add a new sensor for PM2.5.

Also make sure we clamp the values for both PM10 and PM2.5 to 0 instead
of empty string -- that squelches a lot of HASS complaints because
sensors of type "measurement" need to be numeric. In order to reflect
that 0 isn't a real measurement, mark the sensors as unavailable when
the reported value(s) are empty strings.

Finally, bump version to 0.3.0.